% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\title{Authorize with GitHub.}
\usage{
auth(app = ro_app, reauth = FALSE)
}
\arguments{
\item{app}{An \code{\link[httr]{oauth_app}} for GitHub.}

\item{reauth}{(logical) Force re-authorization?}
}
\description{
This function is run automatically to allow gistr to access your GitHub account.
}
\details{
There are two ways to authorise gistr to work with your GitHub account:
\itemize{
 \item Generate a personal access token at
   \url{https://help.github.com/articles/creating-an-access-token-for-command-line-use} and
   record in the \code{GITHUB_PAT} envar.
 \item Interactively login into your GitHub account and authorise with OAuth.
}

Using \code{GITHUB_PAT} is recommended.
}
\examples{
\dontrun{
auth()
auth(reauth = TRUE)
}
}

