% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/issue.R
\name{issue}
\alias{issue}
\title{Open an issue}
\usage{
issue(title, body = NULL, ...)
}
\arguments{
\item{title}{(character) Required. Title of the issue}

\item{body}{(character) Optional. Body of the issue. Use newlines to make line breaks.}

\item{...}{Curl options passed to \code{\link[httr]{POST}}}
}
\description{
Open an issue
}
\examples{
\dontrun{
issue(title = "hello, world!", body = "adflaj asfljas flasjf adfs")
issue(title = "hello there!")
issue(title = "hello there!", body = "adsfasdfdsf \\n asdfadfasdfasdf")
}
}

