% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/propose.R
\name{propose}
\alias{propose}
\title{Propose a package for rOpenSci}
\usage{
propose(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link[devtools]{as.package}} for more information}
}
\description{
Propose a package for rOpenSci
}
\examples{
\dontrun{
propose()
}
}

